refactor: Move direction to new transit model
