+++ src/main/java/org/opentripplanner/gtfs/mapping/TripMapper.java
@@ -3,8 +3,8 @@ package org.opentripplanner.gtfs.mapping;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
-import org.opentripplanner.model.Direction;
 import org.opentripplanner.transit.model.basic.WheelchairAccessibility;
+import org.opentripplanner.transit.model.timetable.Direction;
 import org.opentripplanner.transit.model.timetable.Trip;
 import org.opentripplanner.util.MapUtils;
 import org.slf4j.Logger;
