+++ src/main/java/org/opentripplanner/transit/model/timetable/Trip.java
@@ -8,7 +8,6 @@ import static org.opentripplanner.util.lang.ObjectUtils.ifNotNull;
 import java.util.Objects;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import org.opentripplanner.model.Direction;
 import org.opentripplanner.model.TripAlteration;
 import org.opentripplanner.transit.model.basic.WheelchairAccessibility;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
